Major refactor of server level Cache object and operations
