+++ internal/server/handler.go
@@ -22,7 +22,6 @@ import (
 
 	pb "github.com/datacommonsorg/mixer/internal/proto"
 	"github.com/datacommonsorg/mixer/internal/server/convert"
-	"github.com/datacommonsorg/mixer/internal/server/data"
 	"github.com/datacommonsorg/mixer/internal/server/place"
 	"github.com/datacommonsorg/mixer/internal/server/placein"
 	"github.com/datacommonsorg/mixer/internal/server/recon"
@@ -189,7 +188,7 @@ func (s *Server) GetPlaceStatVars(
 func (s *Server) GetEntityStatVarsUnionV1(
 	ctx context.Context, in *pb.GetEntityStatVarsUnionRequest,
 ) (*pb.GetEntityStatVarsUnionResponse, error) {
-	return statvar.GetEntityStatVarsUnionV1(ctx, in, s.store, s.cache)
+	return statvar.GetEntityStatVarsUnionV1(ctx, in, s.store, s.cachedata)
 }
 
 // GetPropertyLabels implements API for Mixer.GetPropertyLabels.
@@ -309,14 +308,16 @@ func (s *Server) BulkFindEntities(
 	return recon.BulkFindEntities(ctx, in, s.store, s.mapsClient)
 }
 
-// Import implements API for Mixer.Import
-func (s *Server) Import(
-	ctx context.Context, in *pb.ImportRequest,
-) (*pb.ImportResponse, error) {
-	cache, err := data.Import(ctx, in, s.store)
-	if err != nil {
+// UpdateCache implements API for Mixer.UpdateCache
+func (s *Server) UpdateCache(
+	ctx context.Context, in *pb.UpdateCacheRequest,
+) (*pb.UpdateCacheResponse, error) {
+	if err := s.cachedata.UpdateSVGCache(ctx, s.store); err != nil {
 		return nil, err
 	}
-	s.cache = cache
-	return &pb.ImportResponse{Success: true}, nil
+	s.cachedata.UpdateStatVarSearchIndex()
+	if err := s.cachedata.UpdateCustomCache(s.store.SQLClient); err != nil {
+		return nil, err
+	}
+	return &pb.UpdateCacheResponse{Success: true}, nil
 }
