+++ internal/server/place/golden/get_place_stat_date_within_place_test.go
@@ -88,7 +88,7 @@ func TestGetPlaceStatDateWithinPlace(t *testing.T) {
 
 	if err := test.TestDriver(
 		"GetPlaceStatDateWithinPlace",
-		&test.TestOption{UseCache: false},
+		&test.TestOption{},
 		testSuite,
 	); err != nil {
 		t.Errorf("TestDriver() = %s", err)
