+++ internal/server/server.go
@@ -28,6 +28,7 @@ import (
 	pubsub "cloud.google.com/go/pubsub"
 	"github.com/datacommonsorg/mixer/internal/parser/mcf"
 	dcpubsub "github.com/datacommonsorg/mixer/internal/pubsub"
+	"github.com/datacommonsorg/mixer/internal/server/cache"
 	"github.com/datacommonsorg/mixer/internal/server/resource"
 	"github.com/datacommonsorg/mixer/internal/store"
 	"github.com/datacommonsorg/mixer/internal/store/bigtable"
@@ -41,7 +42,7 @@ import (
 type Server struct {
 	store      *store.Store
 	metadata   *resource.Metadata
-	cache      *resource.Cache
+	cachedata  *cache.Cache
 	mapsClient *maps.Client
 	httpClient *http.Client
 }
@@ -143,13 +144,13 @@ func (s *Server) SubscribeBranchCacheUpdate(ctx context.Context) error {
 func NewMixerServer(
 	store *store.Store,
 	metadata *resource.Metadata,
-	cache *resource.Cache,
+	cachedata *cache.Cache,
 	mapsClient *maps.Client,
 ) *Server {
 	return &Server{
 		store:      store,
 		metadata:   metadata,
-		cache:      cache,
+		cachedata:  cachedata,
 		mapsClient: mapsClient,
 		httpClient: &http.Client{},
 	}
