+++ internal/server/statvar/hierarchy/statvar_hierarchy_util.go
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package statvar
+package hierarchy
 
 import (
-	"context"
 	"encoding/json"
 	"os"
 	"path"
@@ -25,11 +24,99 @@ import (
 	"time"
 
 	"github.com/datacommonsorg/mixer/internal/server/resource"
-	"github.com/datacommonsorg/mixer/internal/store"
 
 	pb "github.com/datacommonsorg/mixer/internal/proto"
 	"github.com/datacommonsorg/mixer/internal/util"
 )
+
+const (
+	// SvgRoot is the root stat var group of the hierarchy. It's a virtual entity
+	// that links to the top level category stat var groups.
+	SvgRoot         = "dc/g/Root"
+	CustomSvgRoot   = "dc/g/Custom_Root"
+	CustomSVGPrefix = "dc/g/Custom_"
+)
+
+// Note this function modifies validSVG inside.
+func markValidSVG(
+	input map[string]*pb.StatVarGroupNode,
+	svgID string,
+	validSVG map[string]struct{},
+) bool {
+	// Already checked
+	if _, ok := validSVG[svgID]; ok {
+		return true
+	}
+	svChildren := input[svgID].ChildStatVars
+	svgChildren := input[svgID].ChildStatVarGroups
+	// If there are non-empty sv chldren, then this svg is valid
+	if len(svChildren) > 0 {
+		validSVG[svgID] = struct{}{}
+		return true
+	}
+	// Recursively check child svg, if there is any valid svg child, then this
+	// is valid too
+	for _, svgChild := range svgChildren {
+		if markValidSVG(input, svgChild.Id, validSVG) {
+			validSVG[svgID] = struct{}{}
+			return true
+		}
+	}
+	return false
+}
+
+// FilterSVG filters StatVarGroups based on given stat vars. This does not modify the input
+// data but create a filtered copy of it.
+func FilterSVG(
+	input map[string]*pb.StatVarGroupNode,
+	statVars []string,
+) map[string]*pb.StatVarGroupNode {
+	result := map[string]*pb.StatVarGroupNode{}
+	// Build set for all the given stat vars as valid stat vars.
+	validSV := map[string]struct{}{}
+	for _, sv := range statVars {
+		validSV[sv] = struct{}{}
+	}
+
+	// Step 1: iterate over stat var group, and only keep stat var children with
+	// valid stat vars.
+	for svgID, svgData := range input {
+		filteredChildSV := []*pb.StatVarGroupNode_ChildSV{}
+		for _, childSV := range svgData.ChildStatVars {
+			if _, ok := validSV[childSV.Id]; ok {
+				filteredChildSV = append(filteredChildSV, childSV)
+			}
+		}
+		result[svgID] = &pb.StatVarGroupNode{
+			ChildStatVars:      filteredChildSV,
+			ChildStatVarGroups: svgData.ChildStatVarGroups,
+		}
+	}
+	// Step 2: recursively check if a stat var group is valid. A stat var group
+	// is valid if it has any descendent stat var group with non-empty stat vars
+
+	// All the svg with valid stat vars.
+	validSVG := map[string]struct{}{}
+	for svgID := range result {
+		markValidSVG(result, svgID, validSVG)
+	}
+
+	// Step3: another iteration to only keep valid svg
+	for svgID, svgData := range result {
+		filteredChildren := []*pb.StatVarGroupNode_ChildSVG{}
+		for _, c := range svgData.ChildStatVarGroups {
+			if _, ok := validSVG[c.Id]; ok {
+				filteredChildren = append(filteredChildren, c)
+			}
+		}
+		result[svgID].ChildStatVarGroups = filteredChildren
+		d := result[svgID]
+		if len(d.ChildStatVars) == 0 && len(d.ChildStatVarGroups) == 0 {
+			delete(result, svgID)
+		}
+	}
+	return result
+}
 
 // RemoveSvg removes the blocked svg from the in-memory entries.
 func RemoveSvg(
@@ -54,17 +141,6 @@ func RemoveSvg(
 		}
 	}
 	delete(svgNodeMap, svg)
-}
-
-// GetRawSvg gets the raw svg mapping.
-func GetRawSvg(ctx context.Context, store *store.Store) (
-	map[string]*pb.StatVarGroupNode, error,
-) {
-	svg, err := GetStatVarGroup(ctx, &pb.GetStatVarGroupRequest{}, store, nil)
-	if err != nil {
-		return nil, err
-	}
-	return svg.StatVarGroups, nil
 }
 
 // BuildParentSvgMap gets the mapping of svg/sv id to the parent svg for that
@@ -148,7 +224,7 @@ func getSynonymMap() map[string][]string {
 func BuildStatVarSearchIndex(
 	rawSvg map[string]*pb.StatVarGroupNode,
 	parentSvg map[string][]string,
-	ignoredSvgIds []string,
+	ignoredSvgIds map[string]struct{},
 ) *resource.SearchIndex {
 	defer util.TimeTrack(time.Now(), "BuildStatVarSearchIndex")
 	// map of token to map of sv/svg id to ranking information.
@@ -158,7 +234,7 @@ func BuildStatVarSearchIndex(
 	}
 	ignoredSVG := map[string]string{}
 	// Exclude svg and sv under miscellaneous from the search index
-	for _, svgID := range ignoredSvgIds {
+	for svgID := range ignoredSvgIds {
 		if svgID == "" {
 			continue
 		}
@@ -186,4 +262,58 @@ func BuildStatVarSearchIndex(
 		}
 	}
 	return searchIndex
+}
+
+// AdjustDescendentSVCount returns all the unique stat var under curID.
+// Recursively mutates DescendentStatVarCount for all nodes in id2Node.
+func AdjustDescendentSVCount(
+	id2Node map[string]*pb.StatVarGroupNode, curID string,
+) map[string]struct{} {
+	curNode, ok := id2Node[curID]
+	if !ok {
+		return nil
+	}
+	descendentSV := map[string]struct{}{}
+	for _, sv := range curNode.ChildStatVars {
+		descendentSV[sv.Id] = struct{}{}
+	}
+	for _, childSVG := range curNode.GetChildStatVarGroups() {
+		childDescendentSV := AdjustDescendentSVCount(id2Node, childSVG.GetId())
+		// Child SVG protos have its own DescendentStatVarCount that must be set.
+		childSVG.DescendentStatVarCount = int32(len(childDescendentSV))
+		for sv := range childDescendentSV {
+			descendentSV[sv] = struct{}{}
+		}
+	}
+	curNode.DescendentStatVarCount = int32(len(descendentSV))
+	return descendentSV
+}
+
+// MergeSVGNodes merges node2's svg and svs into node1.
+// Merge here refers to set operation.
+// Warning: node1's DescendentStatVarCount is unchanged.
+// Caller is responsible for calling fixCustomRootDescendentStatVarCount
+// at the end of all merges.
+func MergeSVGNodes(node1, node2 *pb.StatVarGroupNode) {
+	node1SVGs := map[string]bool{}
+	for _, childSVG := range node1.GetChildStatVarGroups() {
+		node1SVGs[childSVG.GetId()] = true
+	}
+	for _, childSVG := range node2.GetChildStatVarGroups() {
+		if _, ok := node1SVGs[childSVG.GetId()]; !ok {
+			node1.ChildStatVarGroups = append(node1.ChildStatVarGroups, childSVG)
+			node1SVGs[childSVG.GetId()] = true
+		}
+	}
+
+	node1SVs := map[string]bool{}
+	for _, childSV := range node1.GetChildStatVars() {
+		node1SVs[childSV.GetId()] = true
+	}
+	for _, childSV := range node2.GetChildStatVars() {
+		if _, ok := node1SVs[childSV.GetId()]; !ok {
+			node1.ChildStatVars = append(node1.ChildStatVars, childSV)
+			node1SVs[childSV.GetId()] = true
+		}
+	}
 }
