+++ internal/server/statvar/search/statvar_search_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package statvar
+package search
 
 import (
 	"testing"
