+++ internal/server/v1/info/golden/variable_group_info_test.go
@@ -114,7 +114,7 @@ func TestVariableGroupInfo(t *testing.T) {
 
 	if err := test.TestDriver(
 		"VariableGroupInfo",
-		&test.TestOption{UseCache: true},
+		&test.TestOption{FetchSVG: true},
 		testSuite,
 	); err != nil {
 		t.Errorf("TestDriver() = %s", err)
