+++ internal/server/v2/observation/golden/existence_test.go
@@ -90,7 +90,7 @@ func TestExistence(t *testing.T) {
 	}
 	if err := test.TestDriver(
 		"Observation(existence)",
-		&test.TestOption{UseCache: true, UseSQLite: true},
+		&test.TestOption{FetchSVG: true, UseSQLite: true},
 		testSuite,
 	); err != nil {
 		t.Errorf("TestDriver() = %s", err)
